refactor PoW context into trait, default to using cuckoo context
