+++ core/src/pow/cuckoo.rs
@@ -17,240 +17,152 @@
 //! simple miner is included, mostly for testing purposes. John Tromp's Tomato
 //! miner will be much faster in almost every environment.
 
+use pow::common::{self, Edge, EdgeType};
+use pow::error::{Error, ErrorKind};
+use pow::num::ToPrimitive;
+use pow::{PoWContext, Proof};
+
 use std::cmp;
 use std::collections::HashSet;
 
-use core::BlockHeader;
-use pow::siphash::siphash24;
-use pow::Proof;
-
 const MAXPATHLEN: usize = 8192;
 
-/// A cuckoo-cycle related error
-#[derive(Debug)]
-pub enum Error {
-	/// Unable to find a short enough path
-	Path,
-	/// Unable to find a solution
-	NoSolution,
-}
-
-/// An edge in the Cuckoo graph, simply references two u64 nodes.
-#[derive(Debug, Copy, Clone, PartialEq, PartialOrd, Eq, Ord, Hash)]
-struct Edge {
-	u: u64,
-	v: u64,
-}
-
 /// Cuckoo cycle context
-pub struct Cuckoo {
-	mask: u64,
-	size: u64,
-	v: [u64; 4],
+pub struct CuckooContext<T>
+where
+	T: EdgeType,
+{
+	edge_bits: u8,
+	siphash_keys: [u64; 4],
+	easiness: T,
+	proof_size: usize,
+	edge_mask: T,
+	graph: Vec<T>,
+	_max_sols: u32,
 }
 
-impl Cuckoo {
-	/// Initializes a new Cuckoo Cycle setup, using the provided byte array to
-	/// generate a seed. In practice for PoW applications the byte array is a
-	/// serialized block header.
-	pub fn from_hash(header_hash: &[u8], sizeshift: u8) -> Cuckoo {
-		println!("header:{:?}", header_hash);
-		let size = 1 << sizeshift;
-		println!("Edge mask: {}", (1 << sizeshift) / 2 - 1);
-		println!("siphash keys [0]: {}", u8_to_u64(header_hash, 0));
-		Cuckoo {
-			v: [
-				u8_to_u64(header_hash, 0),
-				u8_to_u64(header_hash, 8),
-				u8_to_u64(header_hash, 16),
-				u8_to_u64(header_hash, 24),
-			],
-			size: size,
-			mask: (1 << sizeshift) / 2 - 1,
-		}
+impl<T> PoWContext<T> for CuckooContext<T>
+where
+	T: EdgeType,
+{
+	fn new(edge_bits: u8, proof_size: usize, easiness_pct: u32, max_sols: u32) -> Result<Box<Self>, Error> {
+		Ok(Box::new(CuckooContext::<T>::new_impl(edge_bits, proof_size, easiness_pct, max_sols)?))
+	}
+
+	fn set_header_nonce(&mut self, header: Vec<u8>, nonce: Option<u32>) -> Result<(), Error> {
+		self.set_header_nonce_impl(header, nonce)
+	}
+
+	fn find_cycles(&mut self) -> Result<Vec<Proof>, Error> {
+		self.find_cycles_impl()
+	}
+
+	fn verify(&self, proof: &Proof) -> Result<(), Error> {
+		self.verify_impl(proof)
+	}
+}
+
+impl<T> CuckooContext<T>
+where
+	T: EdgeType,
+{
+	/// Create a new cuckoo context with given parameters
+	pub fn new_impl(
+		edge_bits: u8,
+		proof_size: usize,
+		easiness_pct: u32,
+		max_sols: u32,
+	) -> Result<CuckooContext<T>, Error> {
+		let num_edges = 1 << edge_bits;
+		let easiness = easiness_pct.to_u64().ok_or(ErrorKind::IntegerCast)? * num_edges / 100;
+		Ok(CuckooContext {
+			edge_bits: edge_bits,
+			siphash_keys: [0; 4],
+			easiness: T::from(easiness).ok_or(ErrorKind::IntegerCast)?,
+			proof_size: proof_size,
+			edge_mask: T::from(num_edges / 2 - 1).ok_or(ErrorKind::IntegerCast)?,
+			graph: vec![T::zero(); num_edges as usize + 1],
+			_max_sols: max_sols,
+		})
+	}
+
+	fn reset(&mut self) -> Result<(), Error> {
+		let num_edges = 1 << self.edge_bits;
+		self.graph = vec![T::zero(); num_edges + 1];
+		Ok(())
+	}
+
+	/// Set the header and optional nonce in the last part of the header
+	/// and create siphash keys
+	pub fn set_header_nonce_impl(&mut self, header: Vec<u8>, nonce: Option<u32>) -> Result<(), Error> {
+		self.siphash_keys = common::set_header_nonce(header, nonce)?;
+		self.reset()?;
+		Ok(())
 	}
 
 	/// Generates a node in the cuckoo graph generated from our seed. A node is
 	/// simply materialized as a u64 from a nonce and an offset (generally 0 or
 	/// 1).
-	fn new_node(&self, nonce: u64, uorv: u64) -> u64 {
-		return ((siphash24(&self.v, 2 * nonce + uorv) & self.mask) << 1) | uorv;
+	fn new_node(&self, edge: T, uorv: u64) -> Result<T, Error> {
+		common::sipnode::<T>(&self.siphash_keys, edge, &self.edge_mask, uorv, true)
 	}
 
 	/// Creates a new edge in the cuckoo graph generated by our seed from a
 	/// nonce. Generates two node coordinates from the nonce and links them
 	/// together.
-	fn new_edge(&self, nonce: u64) -> Edge {
-		Edge {
-			u: self.new_node(nonce, 0),
-			v: self.new_node(nonce, 1),
-		}
+	fn new_edge(&self, nonce: T) -> Result<Edge<T>, Error> {
+		Ok(Edge {
+			u: self.new_node(nonce, 0)?,
+			v: self.new_node(nonce, 1)?,
+		})
 	}
 
-	/// Assuming increasing nonces all smaller than easiness, verifies the
-	/// nonces form a cycle in a Cuckoo graph. Each nonce generates an edge, we
-	/// build the nodes on both side of that edge and count the connections.
-	pub fn verify(&self, proof: &Proof, ease: u64) -> bool {
-		let easiness = ease * (self.size as u64) / 100;
-		let nonces = &proof.nonces;
-		let mut us = vec![0; proof.proof_size()];
-		let mut vs = vec![0; proof.proof_size()];
-		for n in 0..proof.proof_size() {
-			if nonces[n] >= easiness || (n != 0 && nonces[n] <= nonces[n - 1]) {
-				return false;
-			}
-			us[n] = self.new_node(nonces[n], 0);
-			vs[n] = self.new_node(nonces[n], 1);
-		}
-		let mut i = 0;
-		let mut count = proof.proof_size();
-		loop {
-			let mut j = i;
-			for k in 0..proof.proof_size() {
-				// find unique other j with same vs[j]
-				if k != i && vs[k] == vs[i] {
-					if j != i {
-						return false;
-					}
-					j = k;
-				}
-			}
-			if j == i {
-				return false;
-			}
-			i = j;
-			for k in 0..proof.proof_size() {
-				// find unique other i with same us[i]
-				if k != j && us[k] == us[j] {
-					if i != j {
-						return false;
-					}
-					i = k;
-				}
-			}
-			if i == j {
-				return false;
-			}
-			count -= 2;
-			if i == 0 {
-				break;
-			}
-		}
-		count == 0
-	}
-}
-
-/// Miner for the Cuckoo Cycle algorithm. While the verifier will work for
-/// graph sizes up to a u64, the miner is limited to u32 to be more memory
-/// compact (so shift <= 32). Non-optimized for now and and so mostly used for
-/// tests, being impractical with sizes greater than 2^22.
-pub struct Miner {
-	easiness: u64,
-	proof_size: usize,
-	cuckoo: Cuckoo,
-	graph: Vec<u32>,
-	sizeshift: u8,
-}
-
-/// What type of cycle we have found?
-enum CycleSol {
-	/// A cycle of the right length is a valid proof.
-	ValidProof(Vec<u32>),
-	/// A cycle of the wrong length is great, but not a proof.
-	InvalidCycle(usize),
-	/// No cycles have been found.
-	NoCycle,
-}
-
-impl Miner {
-	/// Creates a new miner for the provided block header
-	pub fn new(header: &BlockHeader, ease: u32, proof_size: usize, sizeshift: u8) -> Miner {
-		Miner::from_hash(header.pre_pow_hash().as_ref(), ease, proof_size, sizeshift)
-	}
-
-	/// Creates a new miner directly from a hash
-	pub fn from_hash(header_hash: &[u8], ease: u32, proof_size: usize, sizeshift: u8) -> Miner {
-		let cuckoo = Cuckoo::from_hash(header_hash, sizeshift);
-		let size = 1 << sizeshift;
-		let graph = vec![0; size + 1];
-		let easiness = (ease as u64) * (size as u64) / 100;
-		println!("Size: {}", size);
-		println!("Easiness: {}", easiness);
-
-		Miner {
-			easiness,
-			cuckoo,
-			graph,
-			proof_size,
-			sizeshift,
-		}
-	}
-
-	/// Searches for a solution
-	pub fn mine(&mut self) -> Result<Proof, Error> {
-		let mut us = [0; MAXPATHLEN];
-		let mut vs = [0; MAXPATHLEN];
-		for nonce in 0..self.easiness {
-			us[0] = self.cuckoo.new_node(nonce, 0) as u32;
-			vs[0] = self.cuckoo.new_node(nonce, 1) as u32;
-			println!("Added us[0], vs[0]: {}, {}", us[0], vs[0]);
-			let u = self.graph[us[0] as usize];
-			let v = self.graph[vs[0] as usize];
-			if us[0] == 0 {
-				continue; // ignore duplicate edges
-			}
-			let nu = self.path(u, &mut us)? as usize;
-			let nv = self.path(v, &mut vs)? as usize;
-
-			let sol = self.find_sol(nu, &us, nv, &vs);
-			match sol {
-				CycleSol::ValidProof(res) => {
-					let mut proof = Proof::new(map_vec!(res.to_vec(), |&n| n as u64));
-					proof.cuckoo_sizeshift = self.sizeshift;
-					return Ok(proof);
-				}
-				CycleSol::InvalidCycle(_) => continue,
-				CycleSol::NoCycle => {
-					self.update_graph(nu, &us, nv, &vs);
-				}
-			}
-		}
-		Err(Error::NoSolution)
-	}
-
-	fn path(&self, mut u: u32, us: &mut [u32]) -> Result<u32, Error> {
+	fn path(&self, mut u: T, us: &mut [T]) -> Result<T, Error> {
 		let mut nu = 0;
-		while u != 0 {
+		while u != T::zero() {
 			nu += 1;
 			if nu >= MAXPATHLEN {
 				while nu != 0 && us[(nu - 1) as usize] != u {
 					nu -= 1;
 				}
-				return Err(Error::Path);
+				return Err(ErrorKind::Path)?;
 			}
 			us[nu as usize] = u;
-			u = self.graph[u as usize];
+			u = self.graph[u.to_u64().ok_or(ErrorKind::IntegerCast)? as usize];
 		}
-		Ok(nu as u32)
+		Ok(T::from(nu).ok_or(ErrorKind::IntegerCast)?)
 	}
 
-	fn update_graph(&mut self, mut nu: usize, us: &[u32], mut nv: usize, vs: &[u32]) {
+	fn update_graph(
+		&mut self,
+		mut nu: usize,
+		us: &[T],
+		mut nv: usize,
+		vs: &[T],
+	) -> Result<(), Error> {
 		if nu < nv {
 			while nu != 0 {
 				nu -= 1;
-				self.graph[us[nu + 1] as usize] = us[nu];
+				self.graph[us[nu + 1].to_u64().ok_or(ErrorKind::IntegerCast)? as usize] = us[nu];
 			}
-			self.graph[us[0] as usize] = vs[0];
+			self.graph[us[0].to_u64().ok_or(ErrorKind::IntegerCast)? as usize] = vs[0];
 		} else {
 			while nv != 0 {
 				nv -= 1;
-				self.graph[vs[nv + 1] as usize] = vs[nv];
+				self.graph[vs[nv + 1].to_u64().ok_or(ErrorKind::IntegerCast)? as usize] = vs[nv];
 			}
-			self.graph[vs[0] as usize] = us[0];
+			self.graph[vs[0].to_u64().ok_or(ErrorKind::IntegerCast)? as usize] = us[0];
 		}
+		Ok(())
 	}
 
-	fn find_sol(&mut self, mut nu: usize, us: &[u32], mut nv: usize, vs: &[u32]) -> CycleSol {
+	fn find_sol(
+		&mut self,
+		mut nu: usize,
+		us: &[T],
+		mut nv: usize,
+		vs: &[T],
+	) -> Result<Vec<T>, Error> {
 		if us[nu] == vs[nv] {
 			let min = cmp::min(nu, nv);
 			nu -= min;
@@ -260,71 +172,148 @@ impl Miner {
 				nv += 1;
 			}
 			if nu + nv + 1 == self.proof_size {
-				self.solution(&us, nu as u32, &vs, nv as u32)
+				self.solution(&us, nu as u64, &vs, nv as u64)
 			} else {
-				CycleSol::InvalidCycle(nu + nv + 1)
+				Err(ErrorKind::InvalidCycle(nu + nv + 1))?
 			}
 		} else {
-			CycleSol::NoCycle
+			Err(ErrorKind::NoCycle)?
 		}
 	}
 
-	fn solution(&mut self, us: &[u32], mut nu: u32, vs: &[u32], mut nv: u32) -> CycleSol {
+	fn solution(&mut self, us: &[T], mut nu: u64, vs: &[T], mut nv: u64) -> Result<Vec<T>, Error> {
 		let mut cycle = HashSet::new();
-		cycle.insert(Edge {
-			u: us[0] as u64,
-			v: vs[0] as u64,
-		});
+		cycle.insert(Edge { u: us[0], v: vs[0] });
 		while nu != 0 {
 			// u's in even position; v's in odd
-			nu -= 1;
+			nu = nu - 1;
 			cycle.insert(Edge {
-				u: us[((nu + 1) & !1) as usize] as u64,
-				v: us[(nu | 1) as usize] as u64,
+				u: us[((nu + 1) & !1) as usize],
+				v: us[(nu | 1) as usize],
 			});
 		}
 		while nv != 0 {
 			// u's in odd position; v's in even
 			nv -= 1;
 			cycle.insert(Edge {
-				u: vs[(nv | 1) as usize] as u64,
-				v: vs[((nv + 1) & !1) as usize] as u64,
+				u: vs[(nv | 1) as usize],
+				v: vs[((nv + 1) & !1) as usize],
 			});
 		}
 		let mut n = 0;
-		let mut sol = vec![0; self.proof_size];
-		for nonce in 0..self.easiness {
-			let edge = self.cuckoo.new_edge(nonce);
+		let mut sol = vec![T::zero(); self.proof_size];
+		for nonce in 0..self.easiness.to_u64().ok_or(ErrorKind::IntegerCast)? as usize {
+			let edge = self.new_edge(T::from(nonce).ok_or(ErrorKind::IntegerCast)?)?;
 			if cycle.contains(&edge) {
-				sol[n] = nonce as u32;
+				sol[n] = T::from(nonce).ok_or(ErrorKind::IntegerCast)?;
 				n += 1;
 				cycle.remove(&edge);
 			}
 		}
 		return if n == self.proof_size {
-			CycleSol::ValidProof(sol)
+			Ok(sol)
 		} else {
-			CycleSol::NoCycle
+			Err(ErrorKind::NoCycle)?
 		};
 	}
-}
 
-/// Utility to transform a 8 bytes of a byte array into a u64.
-fn u8_to_u64(p: &[u8], i: usize) -> u64 {
-	(p[i] as u64)
-		| (p[i + 1] as u64) << 8
-		| (p[i + 2] as u64) << 16
-		| (p[i + 3] as u64) << 24
-		| (p[i + 4] as u64) << 32
-		| (p[i + 5] as u64) << 40
-		| (p[i + 6] as u64) << 48
-		| (p[i + 7] as u64) << 56
+	/// Searches for a solution (simple implementation)
+	pub fn find_cycles_impl(&mut self) -> Result<Vec<Proof>, Error> {
+		let mut us = [T::zero(); MAXPATHLEN];
+		let mut vs = [T::zero(); MAXPATHLEN];
+		for nonce in 0..self.easiness.to_u64().ok_or(ErrorKind::IntegerCast)? as usize {
+			us[0] = self.new_node(T::from(nonce).ok_or(ErrorKind::IntegerCast)?, 0)?;
+			vs[0] = self.new_node(T::from(nonce).ok_or(ErrorKind::IntegerCast)?, 1)?;
+			let u = self.graph[us[0].to_u64().ok_or(ErrorKind::IntegerCast)? as usize];
+			let v = self.graph[vs[0].to_u64().ok_or(ErrorKind::IntegerCast)? as usize];
+			if us[0] == T::zero() {
+				continue; // ignore duplicate edges
+			}
+			let nu = self
+				.path(u, &mut us)?
+				.to_u64()
+				.ok_or(ErrorKind::IntegerCast)? as usize;
+			let nv = self
+				.path(v, &mut vs)?
+				.to_u64()
+				.ok_or(ErrorKind::IntegerCast)? as usize;
+
+			let sol = self.find_sol(nu, &us, nv, &vs);
+			match sol {
+				Ok(s) => {
+					let mut proof = Proof::new(map_vec!(s.to_vec(), |&n| n.to_u64().unwrap_or(0)));
+					proof.cuckoo_sizeshift = self.edge_bits;
+					return Ok(vec![proof]);
+				}
+				Err(e) => match e.kind() {
+					ErrorKind::InvalidCycle(_) => continue,
+					ErrorKind::NoCycle => self.update_graph(nu, &us, nv, &vs)?,
+					_ => return Err(e),
+				},
+			}
+		}
+		Err(ErrorKind::NoSolution)?
+	}
+
+	/// Assuming increasing nonces all smaller than easiness, verifies the
+	/// nonces form a cycle in a Cuckoo graph. Each nonce generates an edge, we
+	/// build the nodes on both side of that edge and count the connections.
+	pub fn verify_impl(&self, proof: &Proof) -> Result<(), Error> {
+		let easiness = self.easiness.to_u64().ok_or(ErrorKind::IntegerCast)?;
+		let nonces = &proof.nonces;
+		let mut us = vec![T::zero(); proof.proof_size()];
+		let mut vs = vec![T::zero(); proof.proof_size()];
+		for n in 0..proof.proof_size() {
+			if nonces[n] >= easiness || (n != 0 && nonces[n] <= nonces[n - 1]) {
+				return Err(ErrorKind::Verification("edge wrong size".to_owned()))?;
+			}
+			us[n] = self.new_node(T::from(nonces[n]).ok_or(ErrorKind::IntegerCast)?, 0)?;
+			vs[n] = self.new_node(T::from(nonces[n]).ok_or(ErrorKind::IntegerCast)?, 1)?;
+		}
+		let mut i = 0;
+		let mut count = proof.proof_size();
+		loop {
+			let mut j = i;
+			for k in 0..proof.proof_size() {
+				// find unique other j with same vs[j]
+				if k != i && vs[k] == vs[i] {
+					if j != i {
+						return Err(ErrorKind::Verification("".to_owned()))?;
+					}
+					j = k;
+				}
+			}
+			if j == i {
+				return Err(ErrorKind::Verification("".to_owned()))?;
+			}
+			i = j;
+			for k in 0..proof.proof_size() {
+				// find unique other i with same us[i]
+				if k != j && us[k] == us[j] {
+					if i != j {
+						return Err(ErrorKind::Verification("".to_owned()))?;
+					}
+					i = k;
+				}
+			}
+			if i == j {
+				return Err(ErrorKind::Verification("".to_owned()))?;
+			}
+			count -= 2;
+			if i == 0 {
+				break;
+			}
+		}
+		match count == 0 {
+			true => Ok(()),
+			false => Err(ErrorKind::Verification("Invalid solution".to_owned()))?,
+		}
+	}
 }
 
 #[cfg(test)]
 mod test {
 	use super::*;
-	use blake2;
 
 	static V1: [u64; 42] = [
 		0x3bbd, 0x4e96, 0x1013b, 0x1172b, 0x1371b, 0x13e6a, 0x1aaa6, 0x1b575, 0x1e237, 0x1ee88,
@@ -357,99 +346,119 @@ mod test {
 		0x701f37b,
 	];
 
-	fn blake2(data: &[u8]) -> blake2::blake2b::Blake2bResult {
-		blake2::blake2b::blake2b(32, &[], data)
+	#[test]
+	fn cuckoo_context() {
+		let ret = mine20_vectors::<u32>();
+		if let Err(r) = ret {
+			panic!("mine20_vectors u32: Error: {}", r);
+		}
+		let ret = mine20_vectors::<u64>();
+		if let Err(r) = ret {
+			panic!("mine20_vectors u64: Error: {}", r);
+		}
+		let ret = validate20_vectors::<u32>();
+		if let Err(r) = ret {
+			panic!("validate20_vectors u32: Error: {}", r);
+		}
+		let ret = validate20_vectors::<u64>();
+		if let Err(r) = ret {
+			panic!("validate20_vectors u64: Error: {}", r);
+		}
+		let ret = validate_fail::<u32>();
+		if let Err(r) = ret {
+			panic!("validate_fail u32: Error: {}", r);
+		}
+		let ret = validate_fail::<u64>();
+		if let Err(r) = ret {
+			panic!("validate_fail u64: Error: {}", r);
+		}
+		let ret = mine16_validate::<u32>();
+		if let Err(r) = ret {
+			panic!("mine16_validate u32: Error: {}", r);
+		}
+		let ret = mine16_validate::<u64>();
+		if let Err(r) = ret {
+			panic!("mine16_validate u64: Error: {}", r);
+		}
 	}
 
 	/// Find a 42-cycle on Cuckoo20 at 75% easiness and verify against a few
 	/// known cycle proofs
 	/// generated by other implementations.
-	#[test]
-	fn mine20_vectors() {
-		let nonces1 = Miner::from_hash(blake2(&[49]).as_bytes(), 75, 42, 20)
-			.mine()
-			.unwrap();
+	fn mine20_vectors<T>() -> Result<(), Error>
+	where
+		T: EdgeType,
+	{
+		let mut cuckoo_ctx = CuckooContext::<T>::new(20, 42, 75, 10)?;
+		cuckoo_ctx.set_header_nonce([49].to_vec(), None)?;
+		let res = cuckoo_ctx.find_cycles()?;
 		let mut proof = Proof::new(V1.to_vec());
 		proof.cuckoo_sizeshift = 20;
-		assert_eq!(proof, nonces1);
+		assert_eq!(proof, res[0]);
 
-		let nonces2 = Miner::from_hash(blake2(&[50]).as_bytes(), 70, 42, 20)
-			.mine()
-			.unwrap();
+		let mut cuckoo_ctx = CuckooContext::<T>::new(20, 42, 70, 10)?;
+		cuckoo_ctx.set_header_nonce([50].to_vec(), None)?;
+		let res = cuckoo_ctx.find_cycles()?;
 		let mut proof = Proof::new(V2.to_vec());
 		proof.cuckoo_sizeshift = 20;
-		assert_eq!(proof, nonces2);
+		assert_eq!(proof, res[0]);
 
-		let nonces3 = Miner::from_hash(blake2(&[51]).as_bytes(), 70, 42, 20)
-			.mine()
-			.unwrap();
+		//re-use context
+		cuckoo_ctx.set_header_nonce([51].to_vec(), None)?;
+		let res = cuckoo_ctx.find_cycles()?;
 		let mut proof = Proof::new(V3.to_vec());
 		proof.cuckoo_sizeshift = 20;
-		assert_eq!(proof, nonces3);
+		assert_eq!(proof, res[0]);
+		Ok(())
 	}
 
-	#[test]
-	fn validate20_vectors() {
-		assert!(
-			Cuckoo::from_hash(blake2(&[49]).as_bytes(), 20)
-				.verify(&Proof::new(V1.to_vec().clone()), 75)
-		);
-		assert!(
-			Cuckoo::from_hash(blake2(&[50]).as_bytes(), 20)
-				.verify(&Proof::new(V2.to_vec().clone()), 70)
-		);
-		assert!(
-			Cuckoo::from_hash(blake2(&[51]).as_bytes(), 20)
-				.verify(&Proof::new(V3.to_vec().clone()), 70)
-		);
+	fn validate20_vectors<T>() -> Result<(), Error>
+	where
+		T: EdgeType,
+	{
+		let mut cuckoo_ctx = CuckooContext::<T>::new(20, 42, 75, 10)?;
+		cuckoo_ctx.set_header_nonce([49].to_vec(), None)?;
+		assert!(cuckoo_ctx.verify(&Proof::new(V1.to_vec().clone())).is_ok());
+		let mut cuckoo_ctx = CuckooContext::<T>::new(20, 42, 70, 10)?;
+		cuckoo_ctx.set_header_nonce([50].to_vec(), None)?;
+		assert!(cuckoo_ctx.verify(&Proof::new(V2.to_vec().clone())).is_ok());
+		cuckoo_ctx.set_header_nonce([51].to_vec(), None)?;
+		assert!(cuckoo_ctx.verify(&Proof::new(V3.to_vec().clone())).is_ok());
+		Ok(())
 	}
 
-	/// Just going to disable this for now, as it's painful to try and get a
-	/// valid cuckoo28 vector (TBD: 30 is more relevant now anyhow)
-	#[test]
-	#[ignore]
-	fn validate28_vectors() {
-		let mut test_header = [0; 32];
-		test_header[0] = 24;
-		assert!(Cuckoo::from_hash(&test_header, 28).verify(&Proof::new(V4.to_vec().clone()), 50));
-	}
-
-	#[test]
-	fn validate_fail() {
+	fn validate_fail<T>() -> Result<(), Error>
+	where
+		T: EdgeType,
+	{
 		// edge checks
-		assert!(
-			!Cuckoo::from_hash(blake2(&[49]).as_bytes(), 20).verify(&Proof::new(vec![0; 42]), 75)
-		);
-		assert!(
-			!Cuckoo::from_hash(blake2(&[49]).as_bytes(), 20)
-				.verify(&Proof::new(vec![0xffff; 42]), 75)
-		);
+		let mut cuckoo_ctx = CuckooContext::<T>::new(20, 42, 75, 10)?;
+		cuckoo_ctx.set_header_nonce([49].to_vec(), None)?;
+		// edge checks
+		assert!(!cuckoo_ctx.verify(&Proof::new(vec![0; 42])).is_ok());
+		assert!(!cuckoo_ctx.verify(&Proof::new(vec![0xffff; 42])).is_ok());
 		// wrong data for proof
-		assert!(
-			!Cuckoo::from_hash(blake2(&[50]).as_bytes(), 20)
-				.verify(&Proof::new(V1.to_vec().clone()), 75)
-		);
+		cuckoo_ctx.set_header_nonce([50].to_vec(), None)?;
+		assert!(!cuckoo_ctx.verify(&Proof::new(V1.to_vec().clone())).is_ok());
 		let mut test_header = [0; 32];
 		test_header[0] = 24;
-		assert!(
-			!Cuckoo::from_hash(blake2(&test_header).as_bytes(), 20)
-				.verify(&Proof::new(V4.to_vec().clone()), 50)
-		);
+		let mut cuckoo_ctx = CuckooContext::<T>::new(20, 42, 50, 10)?;
+		cuckoo_ctx.set_header_nonce(test_header.to_vec(), None)?;
+		assert!(!cuckoo_ctx.verify(&Proof::new(V4.to_vec().clone())).is_ok());
+		Ok(())
 	}
 
-	#[test]
-	fn mine20_validate() {
-		// cuckoo20
+	fn mine16_validate<T>() -> Result<(), Error>
+	where
+		T: EdgeType,
+	{
 		for n in 1..5 {
 			let h = [n; 32];
-			let nonces = Miner::from_hash(&h, 75, 42, 20).mine().unwrap();
-			assert!(Cuckoo::from_hash(&h, 20).verify(&nonces, 75));
+			let mut cuckoo_ctx = CuckooContext::<T>::new(16, 42, 75, 10)?;
+			cuckoo_ctx.set_header_nonce(h.to_vec(), None)?;
+			let res = cuckoo_ctx.find_cycles()?;
+			assert!(cuckoo_ctx.verify(&res[0]).is_ok())
 		}
-		// cuckoo18
-		for n in 1..5 {
-			let h = [n; 32];
-			let nonces = Miner::from_hash(&h, 75, 42, 18).mine().unwrap();
-			assert!(Cuckoo::from_hash(&h, 18).verify(&nonces, 75));
-		}
+		Ok(())
 	}
 }
