--- core/src/pow/miner.rs
@@ -1,1 +1,0 @@
-
