+++ core/src/pow/mod.rs
@@ -39,7 +39,6 @@ extern crate grin_util as util;
 mod common;
 pub mod cuckatoo;
 pub mod cuckoo;
-pub mod cuckoo_ctx;
 mod error;
 mod siphash;
 mod types;
@@ -49,15 +48,21 @@ use consensus;
 use core::{Block, BlockHeader};
 use genesis;
 use global;
-use pow::cuckoo::{Cuckoo, Error};
 
+pub use pow::cuckoo::CuckooContext;
+pub use pow::cuckatoo::CuckatooContext;
 pub use self::types::*;
+pub use self::common::EdgeType;
+pub use pow::error::Error;
+
+const MAX_SOLS:u32 = 10;
 
 /// Validates the proof of work of a given header, and that the proof of work
 /// satisfies the requirements of the header.
-pub fn verify_size(bh: &BlockHeader, cuckoo_sz: u8) -> bool {
-	Cuckoo::from_hash(bh.pre_pow_hash().as_ref(), cuckoo_sz)
-		.verify(&bh.pow.proof, consensus::EASINESS as u64)
+pub fn verify_size(bh: &BlockHeader, cuckoo_sz: u8) -> Result<(), Error> {
+	let mut ctx = global::create_pow_context::<u64>(cuckoo_sz, bh.pow.proof.nonces.len(), consensus::EASINESS, MAX_SOLS)?;
+	ctx.set_header_nonce(bh.pre_pow_hash().to_vec(), None)?;
+	ctx.verify(&bh.pow.proof)
 }
 
 /// Mines a genesis block using the internal miner
@@ -74,7 +79,7 @@ pub fn mine_genesis_block() -> Result<Block, Error> {
 	let sz = global::min_sizeshift();
 	let proof_size = global::proofsize();
 
-	pow_size(&mut gen.header, genesis_difficulty, proof_size, sz).unwrap();
+	pow_size(&mut gen.header, genesis_difficulty, proof_size, sz)?;
 	Ok(gen)
 }
 
@@ -86,7 +91,8 @@ pub fn pow_size(
 	diff: Difficulty,
 	proof_size: usize,
 	sz: u8,
-) -> Result<(), Error> {
+) -> Result<(), Error>
+{
 	let start_nonce = bh.pow.nonce;
 
 	// set the nonce for faster solution finding in user testing
@@ -98,8 +104,10 @@ pub fn pow_size(
 	loop {
 		// if we found a cycle (not guaranteed) and the proof hash is higher that the
 		// diff, we're all good
-		if let Ok(proof) = cuckoo::Miner::new(bh, consensus::EASINESS, proof_size, sz).mine() {
-			bh.pow.proof = proof;
+		let mut ctx = global::create_pow_context::<u32>(sz, proof_size, consensus::EASINESS, MAX_SOLS)?;
+		ctx.set_header_nonce(bh.pre_pow_hash().to_vec(), None)?;
+		if let Ok(proofs) = ctx.find_cycles() {
+			bh.pow.proof = proofs[0].clone();
 			if bh.pow.to_difficulty() >= diff {
 				return Ok(());
 			}
@@ -137,6 +145,6 @@ mod test {
 		).unwrap();
 		assert!(b.header.pow.nonce != 310);
 		assert!(b.header.pow.to_difficulty() >= Difficulty::one());
-		assert!(verify_size(&b.header, global::min_sizeshift()));
+		assert!(verify_size(&b.header, global::min_sizeshift()).is_ok());
 	}
 }
